pyproject.toml
[project]
name = "chachich-spotify-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "python-dotenv",
    "langchain-groq",
    "langgraph",
    "mcp-use",
]
